bvh-bvh: remove virtualisation in `collide` call of BVHCollide
