+++ src/collision_node.cpp
@@ -35,56 +35,5 @@
 
 /** \author Jia Pan */
 
-#include <../src/collision_node.h>
-#include "coal/internal/traversal_recurse.h"
-
-namespace coal {
-
-void checkResultLowerBound(const CollisionResult& result,
-                           Scalar sqrDistLowerBound) {
-  COAL_UNUSED_VARIABLE(result);
-  const Scalar dummy_precision = std::sqrt(Eigen::NumTraits<Scalar>::epsilon());
-  COAL_UNUSED_VARIABLE(dummy_precision);
-  if (sqrDistLowerBound == 0) {
-    COAL_ASSERT(result.distance_lower_bound <= dummy_precision,
-                "Distance lower bound should not be positive.",
-                std::logic_error);
-  } else {
-    COAL_ASSERT(result.distance_lower_bound * result.distance_lower_bound -
-                        sqrDistLowerBound <
-                    dummy_precision,
-                "Distance lower bound and sqrDistLowerBound should coincide.",
-                std::logic_error);
-  }
-}
-
-void collide(CollisionTraversalNodeBase* node, const CollisionRequest& request,
-             CollisionResult& result, BVHFrontList* front_list,
-             bool recursive) {
-  if (front_list && front_list->size() > 0) {
-    propagateBVHFrontListCollisionRecurse(node, request, result, front_list);
-  } else {
-    Scalar sqrDistLowerBound = 0;
-    if (recursive)
-      collisionRecurse(node, 0, 0, front_list, sqrDistLowerBound);
-    else
-      collisionNonRecurse(node, front_list, sqrDistLowerBound);
-    if (!std::isnan(sqrDistLowerBound)) {
-      checkResultLowerBound(result, sqrDistLowerBound);
-    }
-  }
-}
-
-void distance(DistanceTraversalNodeBase* node, BVHFrontList* front_list,
-              unsigned int qsize) {
-  node->preprocess();
-
-  if (qsize <= 2)
-    distanceRecurse(node, 0, 0, front_list);
-  else
-    distanceQueueRecurse(node, 0, 0, front_list, qsize);
-
-  node->postprocess();
-}
-
-}  // namespace coal
+// collide() and distance() are now inline templates in collision_node.h.
+// This file is intentionally empty.
